requirements.txt
fastapi==0.115.14
google-generativeai==0.8.3
googletrans-py==4.0.0
markitdown==0.1.2
protobuf>=3.20.2,<6.0.0
pydantic==2.11.7
python-dotenv==1.1.1
supabase>=1.0.4
uvicorn==0.34.3
